Refactor after UnitTestnig

+++ src/Footballize/Footballize.Services/Data/ProvinceService.cs
@@ -31,8 +31,8 @@
             return this.provincesRepository
                 .All()
                 .Where(p => p.CountryId.Equals(countryId))
-                .OrderBy(p=>p.Name)
-                .ThenByDescending(p=>p.Towns.Count)
+                .OrderBy(p => p.Name)
+                .ThenByDescending(p => p.Towns.Count)
                 .To<TViewModel>();
         }
 
@@ -51,6 +51,12 @@
         public async Task RemoveProvinceAsync(string id)
         {
             var provinceToRemove = await this.provincesRepository.GetByIdAsync(id);
+
+            if (provinceToRemove == null)
+            {
+                throw new ServiceException(string.Format(GlobalConstants.EntityCannotBeNullErrorMessage, nameof(Province)));
+            }
+
             this.provincesRepository.Delete(provinceToRemove);
             await this.provincesRepository.SaveChangesAsync();
         }
